worker/requirements.txt
kafka-python
redis
pandas
scikit-learn
xgboost
circuitbreaker
lz4